Extract fs wrapper to util file
