+++ apis/python/src/tiledbsoma/io/ingest.py
@@ -14,7 +14,6 @@ import math
 import time
 from typing import (
     Any,
-    ContextManager,
     Dict,
     List,
     Mapping,
@@ -27,7 +26,6 @@ from typing import (
     cast,
     overload,
 )
-from unittest import mock
 
 import anndata as ad
 import h5py
@@ -36,7 +34,6 @@ import pandas as pd
 import pyarrow as pa
 import scipy.sparse as sp
 import tiledb
-from anndata._core import file_backing
 from anndata._core.sparse_dataset import SparseDataset
 from somacore.options import PlatformConfig
 from typing_extensions import get_args
@@ -84,6 +81,7 @@ from ._registration import (
     get_dataframe_values,
     signatures,
 )
+from ._util import read_h5ad
 
 _NDArr = TypeVar("_NDArr", bound=NDArray)
 _TDBO = TypeVar("_TDBO", bound=TileDBObject[RawHandle])
@@ -199,42 +197,6 @@ def register_anndatas(
     )
 
 
-# This trick lets us ingest H5AD with "r" (backed mode) from S3 URIs.  While h5ad
-# supports any file-like object, AnnData specifically wants only an `os.PathLike`
-# object. The only thing it does with the PathLike is to use it to get the filename.
-class _FSPathWrapper:
-    """Tricks anndata into thinking a file-like object is an ``os.PathLike``.
-
-    While h5ad supports any file-like object, anndata specifically wants
-    an ``os.PathLike object``, which it uses *exclusively* to get the "filename"
-    of the opened file.
-
-    We need to provide ``__fspath__`` as a real class method, so simply
-    setting ``some_file_obj.__fspath__ = lambda: "some/path"`` won't work,
-    so here we just proxy all attributes except ``__fspath__``.
-    """
-
-    def __init__(self, obj: object, path: Path) -> None:
-        self._obj = obj
-        self._path = path
-
-    def __fspath__(self) -> Path:
-        return self._path
-
-    def __getattr__(self, name: str) -> object:
-        return getattr(self._obj, name)
-
-
-def _hack_patch_anndata() -> ContextManager[object]:
-    """Part Two of the ``_FSPathWrapper`` trick."""
-
-    @file_backing.AnnDataFileManager.filename.setter  # type: ignore
-    def filename(self: file_backing.AnnDataFileManager, filename: Path) -> None:
-        self._filename = filename
-
-    return mock.patch.object(file_backing.AnnDataFileManager, "filename", filename)
-
-
 def from_h5ad(
     experiment_uri: str,
     input_path: Path,
@@ -346,27 +308,33 @@ def from_h5ad(
     s = _util.get_start_stamp()
     logging.log_io(None, f"START  Experiment.from_h5ad {input_path}")
 
-    with tiledb.VFS(ctx=context.tiledb_ctx).open(input_path) as input_handle:
-        logging.log_io(None, f"START  READING {input_path}")
-        with _hack_patch_anndata():
-            anndata = ad.read_h5ad(_FSPathWrapper(input_handle, input_path), "r")
-        logging.log_io(None, _util.format_elapsed(s, f"FINISH READING {input_path}"))
-        uri = from_anndata(
-            experiment_uri,
-            anndata,
-            measurement_name,
-            context=context,
-            platform_config=platform_config,
-            obs_id_name=obs_id_name,
-            var_id_name=var_id_name,
-            X_layer_name=X_layer_name,
-            raw_X_layer_name=raw_X_layer_name,
-            ingest_mode=ingest_mode,
-            use_relative_uri=use_relative_uri,
-            X_kind=X_kind,
-            registration_mapping=registration_mapping,
-            uns_keys=uns_keys,
-        )
+    logging.log_io(None, f"START  READING {input_path}")
+
+    # The input handle needs to remain open while we process the file-backed
+    # AnnData object.
+    input_handle, anndata = read_h5ad(input_path, mode="r", ctx=context.tiledb_ctx)
+    logging.log_io(None, _util.format_elapsed(s, f"FINISH READING {input_path}"))
+
+    uri = from_anndata(
+        experiment_uri,
+        anndata,
+        measurement_name,
+        context=context,
+        platform_config=platform_config,
+        obs_id_name=obs_id_name,
+        var_id_name=var_id_name,
+        X_layer_name=X_layer_name,
+        raw_X_layer_name=raw_X_layer_name,
+        ingest_mode=ingest_mode,
+        use_relative_uri=use_relative_uri,
+        X_kind=X_kind,
+        registration_mapping=registration_mapping,
+        uns_keys=uns_keys,
+    )
+
+    # Close the input handle now that we're done processinig the file-backed
+    # AnnData object.
+    input_handle.close()
 
     logging.log_io(
         None, _util.format_elapsed(s, f"FINISH Experiment.from_h5ad {input_path} {uri}")
